Актуализировал эндпоинты, добавил DTO

+++ CarWashService/src/main/java/ru/edu/repository/TimeTableRepository.java
@@ -32,7 +32,7 @@ public interface TimeTableRepository extends JpaRepository<TimeTable, TimeTableI
 
     @Query(value = "delete from time_table tt " +
             "where date_trunc('minute',tt.date_table) = date_trunc('minute',':date') and tt.car_wash_id = :carWashId and tt.user_info_id_user = :idUser", nativeQuery=true)
-    TimeTable abandonOrder(@Param("date")Date date, @Param("carWashId")Long carWashId, @Param("idUser")Integer idUser);
+    TimeTable abandonOrder(@Param("date")Date date, @Param("carWashId")Long carWashId, @Param("idUser")Long idUser);
 
 
 }
